Split `Alert` and `PlannedTest` models
They don’t have much in common, so we don’t gain a lot from making them
share a module or inherit from each other.
