+++ app/models/alert.py
@@ -6,21 +6,12 @@ from notifications_utils.serialised_model import SerialisedModel
 from app.models.alert_date import AlertDate
 
 
-class BaseAlert(SerialisedModel):
+class Alert(SerialisedModel):
     ALLOWED_PROPERTIES = {
-        'starts_at',
-    }
-
-    @property
-    def starts_at_date(self):
-        return AlertDate(self.starts_at)
-
-
-class Alert(BaseAlert):
-    ALLOWED_PROPERTIES = BaseAlert.ALLOWED_PROPERTIES | {
         'id',
         'channel',
         'approved_at',
+        'starts_at',
         'cancelled_at',
         'finishes_at',
         'areas',
@@ -46,6 +37,10 @@ class Alert(BaseAlert):
     @property
     def approved_at_date(self):
         return AlertDate(self.approved_at)
+
+    @property
+    def starts_at_date(self):
+        return AlertDate(self.starts_at)
 
     @property
     def expires_date(self):
@@ -81,13 +76,3 @@ class Alert(BaseAlert):
             self.expires_date.as_utc_datetime >= now and
             self.approved_at_date.as_utc_datetime <= now
         )
-
-
-class PlannedTest(BaseAlert):
-    ALLOWED_PROPERTIES = BaseAlert.ALLOWED_PROPERTIES | {
-        'display_areas',
-        'description',
-    }
-
-    def __lt__(self, other):
-        return self.starts_at < other.starts_at
