+++ tests/app/main/views/test_planned_tests.py
@@ -2,8 +2,8 @@ import pytest
 from dateutil.parser import parse as dt_parse
 from freezegun import freeze_time
 
-from app.models.alert import PlannedTest
 from app.models.alerts import Alerts
+from app.models.planned_test import PlannedTest
 from tests import normalize_spaces
 from tests.conftest import create_alert_dict
